tests: moved expectThrowsAsync to utils
